requirements.txt
numpy
langchain
pymilvus
python-dotenv
cohere
docling
cryptography
langchain
langchain-text-splitters
phe
pycryptodome
matplotlib